requirements.txt
pandas==2.0.3
numpy==1.24.3
pyyaml==6.0
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.65.0
colorama==0.4.6
pathlib==1.0.1
